+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-## June 30 2021 - v.34.0
-    - Exteneded 
-    
+## March 26 2021 - v.34.1
+    - Exteneded fileset info
+
 ## June 30 2020 - v.33.0
     - Added sampler